Use user defined hover text for images in tasks
With changing the image hover handling when metadata were introduced,
static images and draggable images used the image's metadata title
for the hover text, but not what the user specified in the
editor of Drag-and-Drop.

This change overrides the image hover text for static images and
draggable images. It will be set to what's specified by the user,
so possibly empty and not displayed at all.

There's a similar pull request for the view of Drag-and-Drop.