JI-2595 Fix dropzone's parameters when disassociating draggable
When disassociating a draggable with a dropzone where the draggable
was correct only the draggable was updated. And thus the dropzone
had "hidden" correct elements that could never be dragged there
making the total score unachievable. This removes the draggables
that can no longer be dropped into a dropzone as that dropzone's
correct elements.